+++ libs/core/src/lib/input-group/input-group.component.ts
@@ -16,8 +16,7 @@ import {
     ViewEncapsulation
 } from '@angular/core';
 import { ControlValueAccessor, NG_VALUE_ACCESSOR } from '@angular/forms';
-import { filter, fromEvent, map, merge, Observable, Subject } from 'rxjs';
-import { takeUntil } from 'rxjs/operators';
+import { filter, fromEvent, map, merge, Observable, Subject, takeUntil, debounceTime } from 'rxjs';
 
 import { FormStates, Nullable } from '@fundamental-ngx/core/shared';
 import { FormItemControl, registerFormItemControl } from '@fundamental-ngx/core/form';
@@ -316,6 +315,8 @@ export class InputGroupComponent implements ControlValueAccessor, AfterViewInit,
             fromEvent(inputElement, 'focusin').pipe(map(() => true)),
             fromEvent(inputElement, 'focusout').pipe(map(() => false))
         ).pipe(
+            // debounceTime is needed in order to filter subsequent focus-blur events, that happen simultaneously
+            debounceTime(10),
             filter(() => this.showFocus),
             takeUntil(this._onDestroy$)
         );