+++ libs/core/src/lib/token/tokenizer.component.spec.ts
@@ -60,9 +60,9 @@ describe('TokenizerComponent', () => {
         expect(component.buildComponentCssClass).toHaveBeenCalled();
     });
 
-    it('should addEventListener to input during ngAfterViewChecked and handle keydown', async () => {
+    it('should addEventListener to input during ngAfterViewInit and handle keydown', async () => {
         spyOn(component, 'handleKeyDown');
-        component.ngAfterViewChecked();
+        component.ngAfterViewInit();
 
         await whenStable(fixture);
 
@@ -115,7 +115,7 @@ describe('TokenizerComponent', () => {
         const event = new MouseEvent('click', {
             ctrlKey: true
         });
-        component.ngAfterViewChecked();
+        component.ngAfterViewInit();
         (component.tokenList.first.elementRef.nativeElement.querySelector('.fd-token') as HTMLElement).dispatchEvent(
             event
         );
@@ -128,7 +128,7 @@ describe('TokenizerComponent', () => {
     });
 
     it('should deselect using control or command', () => {
-        component.ngAfterViewChecked();
+        component.ngAfterViewInit();
         const event = new MouseEvent('click', {
             ctrlKey: true
         });
@@ -147,7 +147,7 @@ describe('TokenizerComponent', () => {
     });
 
     it('should select using shift', () => {
-        component.ngAfterViewChecked();
+        component.ngAfterViewInit();
         const event = new MouseEvent('click', {
             ctrlKey: false,
             shiftKey: true
@@ -163,10 +163,6 @@ describe('TokenizerComponent', () => {
         component.tokenList.forEach((token) =>
             spyOn(token.elementRef.nativeElement.querySelector('.fd-token'), 'focus')
         );
-        component.tokenList.forEach((token) =>
-            spyOn(token.elementRef.nativeElement.querySelector('.fd-token'), 'setAttribute')
-        );
-        spyOn(component, 'addKeyboardListener');
         spyOn(component, 'handleKeyDown');
 
         component.focusTokenElement(1);
@@ -178,25 +174,6 @@ describe('TokenizerComponent', () => {
             .filter((element, index) => index === 1)[0]
             .elementRef.nativeElement.querySelector('.fd-token');
         expect(elementToCheck.focus).toHaveBeenCalled();
-        expect(elementToCheck.setAttribute).toHaveBeenCalledWith('tabindex', '0');
-        expect(component.addKeyboardListener).toHaveBeenCalledWith(elementToCheck, 1);
-    });
-
-    it('should add keyboard listener', async () => {
-        spyOn(component, 'handleKeyDown');
-        const mockElement = document.createElement('span');
-        spyOn(mockElement, 'addEventListener').and.callThrough();
-        spyOn(mockElement, 'setAttribute');
-        spyOn(mockElement, 'removeEventListener');
-        const event = new KeyboardEvent('blur');
-        component.addKeyboardListener(mockElement, 0);
-        mockElement.dispatchEvent(event);
-
-        await whenStable(fixture);
-
-        expect(mockElement.addEventListener).toHaveBeenCalled();
-        expect(mockElement.setAttribute).toHaveBeenCalled();
-        expect(mockElement.removeEventListener).toHaveBeenCalled();
     });
 
     it('should handle resize - getting smaller', () => {
@@ -277,7 +254,7 @@ describe('TokenizerComponent', () => {
         });
         spyOnProperty(component.tokenizerInnerEl.nativeElement, 'scrollWidth').and.returnValue(5);
 
-        component.ngAfterViewChecked();
+        component.ngAfterViewInit();
 
         await whenStable(fixture);