+++ libs/fn/src/lib/cdk/toast/interfaces/toast-duration-dismissible-container-component.interface.ts
@@ -1,0 +1,7 @@
+import { ToastContainerComponent } from './toast-container-component.interface';
+import { Subject } from 'rxjs';
+
+export interface ToastDurationDismissibleContainerComponent<P> extends ToastContainerComponent<P> {
+    mouseover$: Subject<void>;
+    mouseleave$: Subject<void>;
+}